tests new

+++ math/s21_math.h
@@ -13,8 +13,8 @@ double s21_cos(double x);
 int s21_abs(int x);
 double s21_fabs(double x);
 
-long double s21_floor(double x);
-long double s21_ceil(double x);
+double s21_floor(double x);
+double s21_ceil(double x);
 
 long double s21_fmod(double x, double y);
 long double s21_log(double x);
